+++ FawrySystemService/src/main/java/com/FawrySystem/FawrySystemService/models/payment/CODPayment.java
@@ -3,7 +3,7 @@ package com.FawrySystem.FawrySystemService.models.payment;
 import com.FawrySystem.FawrySystemService.models.Users.Customer;
 
 public class CODPayment implements PaymentStrategy {
-    public boolean pay(double amount,  Customer customer){
+    public boolean pay(float amount,  Customer customer){
         return true;
     }
