+++ FawrySystemService/src/main/java/com/FawrySystem/FawrySystemService/models/payment/PaymentStrategy.java
@@ -3,5 +3,5 @@ package com.FawrySystem.FawrySystemService.models.payment;
 import com.FawrySystem.FawrySystemService.models.Users.Customer;
 
 public interface PaymentStrategy {
-    public boolean pay(double amount, Customer customer);
+    public boolean pay(float amount, Customer customer);
 }
